+++ src/runtimes/index.js
@@ -1,0 +1,9 @@
+const { zipGoFunction } = require('./go')
+const { zipJsFunction } = require('./node')
+const { zipRustFunction } = require('./rust')
+
+module.exports = {
+  go: zipGoFunction,
+  js: zipJsFunction,
+  rust: zipRustFunction,
+}
